# ENH: Improve directional diagram algo