## Description
This branch changes the way we generate directional diagrams from partial diagrams. Since the paths for the directional diagrams are known ahead of time, we can use a depth-first-search algorithm to find the directed versions of the paths. `NetworkX.dfs_tree` returns the correct diagrams but with the edges reversed (our previous algorithm also did this) so we simply reverse the edges like before and voila! The new algorithm is significantly more efficient for larger graphs (see results below) and still noticeably faster for small graphs. 

## Changes

* Changes directional diagram algorithm 
to use depth-first-search (i.e. `nx.dfs_tree`)
to create the directional diagrams. Also
uses the `nx.DiGraph.reverse` method to build
the directional diagram with reversed edges
and changes directional diagram return type
for `return_edges=False` to a `nx.DiGraph`.

* Removes private functions 
`diagrams._collect_sources` and
`diagrams.get_directional_path_edges`

* Move array-based edge flipping code into
the `return_edges=True` code path

* Addresses the directional diagram 
algo improvement portion of #22

# Performance changes

On this branch we see a significant increase in performance for many of the benchmarks. Running `$ asv continuous master improve_dir_edge_generation` showed improvements even for some of the functions from `calculations.py` because they rely on generating the diagrams to create the expressions.
```
| Change   | Before [9493049b] <master>   | After [0ae72b58] <improve_dir_edge_generation>   |   Ratio | Benchmark (Parameter)
|----------|------------------------------|--------------------------------------------------|---------|--------------------------------------------------------|
| -        | 3.17±0.3ms                   | 2.22±0.1ms                                       |    0.7  | time_calc_state_probs('3-state', True)
| -        | 83.1±10μs                    | 54.3±2μs                                         |    0.65 | time_calc_sigma('3-state', False)
| -        | 14.4±0.9ms                   | 8.73±0.1ms                                       |    0.61 | time_calc_state_probs('Hill-5-state', True)
| -        | 1.57±0.03ms                  | 807±20μs                                         |    0.51 | time_generate_directional_diagrams('3-state', False)
| -        | 687±40ms                     | 352±2ms                                          |    0.51 | time_generate_directional_diagrams('EmrE-8-state', False)
| -        | 1.47±0.07ms                  | 715±20μs                                         |    0.49 | time_calc_state_probs('3-state', False)
| -        | 9.88±0.2ms                   | 4.89±0.3ms                                       |    0.49 | time_generate_directional_diagrams('Hill-5-state', False)
| -        | 102±0.8ms                    | 49.8±1ms                                         |    0.49 | time_generate_directional_diagrams('Hill-8-state', False)
| -        | 1.29±0.02ms                  | 584±30μs                                         |    0.45 | time_generate_directional_diagrams('3-state', True)
| -        | 495±6ms                      | 199±10ms                                         |    0.4  | time_calc_state_probs('EmrE-8-state', False)
| -        | 8.10±0.3ms                   | 3.23±0.02ms                                      |    0.4  | time_calc_state_probs('Hill-5-state', False)
| -        | 74.6±2ms                     | 28.4±1ms                                         |    0.38 | time_calc_state_probs('Hill-8-state', False)
| -        | 471±4ms                      | 165±3ms                                          |    0.35 | time_generate_directional_diagrams('EmrE-8-state', True)
| -        | 7.38±0.09ms                  | 2.53±0.03ms                                      |    0.34 | time_generate_directional_diagrams('Hill-5-state', True)
| -        | 80.8±20ms                    | 23.2±0.3ms                                       |    0.29 | time_generate_directional_diagrams('Hill-8-state', True)

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```