@nawtrey 2024-08-17
Just as a final performance check, I went ahead and added a new test case for the diagram benchmarks which tests a maximally-connected 7-state diagram. For maximally-connected graphs, there are `N^(N-1)` directional diagrams, and thus `117649` directional diagrams for a maximally-connected 7-state diagram. For reference, the 8-state EmrE model has `3072` directional diagrams.

Here is the performance comparison with the new test model included for the latest commit:
```
[58.33%] ··· ...lDiagrams.time_generate_directional_diagrams                ok
[58.33%] ··· ============== ============ ============
             --                    return_edges
             -------------- -------------------------
                 graph          True        False
             ============== ============ ============
                3-state      623±100μs    791±100μs
              Hill-5-state   2.82±0.2ms   4.14±0.4ms
              Hill-8-state   23.0±0.4ms    46.7±2ms
              EmrE-8-state    179±8ms      328±20ms
              Max-7-state    8.28±0.04s   12.7±0.07s
             ============== ============ ============
```  
And for `master`:
```
[83.33%] ··· ...lDiagrams.time_generate_directional_diagrams                ok
[83.33%] ··· ============== ============ ============
             --                    return_edges
             -------------- -------------------------
                 graph          True        False
             ============== ============ ============
                3-state      1.31±0.1ms   2.49±0.3ms
              Hill-5-state    7.64±2ms     11.4±2ms
              Hill-8-state   80.3±10ms     104±5ms
              EmrE-8-state    478±70ms    785±200ms
              Max-7-state    19.9±0.9s    25.4±0.1s
             ============== ============ ============
```
It appears for diagrams this complex there is also a benefit for peak memory (probably due to the use of the less complex `nx.DiGraph` objects):
```
| Change   | Before [9493049b] <master>   | After [5a5a77b0] <improve_dir_edge_generation>   |   Ratio | Benchmark (Parameter)
|----------|------------------------------|--------------------------------------------------|---------|----------------------------------
| -        | 1.05G                        | 818M                                             |    0.78 | peakmem_generate_directional_diagrams('Max-7-state', False)
```

I don't think I will be adding the 7-state test case to the benchmark suite due to the long run time of `12.7` seconds, but I think it is useful to see what kind of performance we get for such an extreme case (i.e. `12.7`s run time with `0.82` GB used). I think it is impressive that I'm able to build ~`9300` directional diagrams per second on my laptop.

For future reference I will paste the benchmark output with the 7-state case included:
<details><summary> Benchmarks ran with Max-7-state</summary>

```
$ asv continuous master improve_dir_edge_generation -b bench_diagrams*
Couldn't load asv.plugins._mamba_helpers because
No module named 'libmambapy'
· Creating environments
· Discovering benchmarks
·· Uninstalling from conda-py3.11-pip+networkx-pip+numpy-pip+pytest-pip+sympy
·· Building 5a5a77b0 <improve_dir_edge_generation> for conda-py3.11-pip+networkx-pip+numpy-pip+pytest-pip+sympy
·· Installing 5a5a77b0 <improve_dir_edge_generation> into conda-py3.11-pip+networkx-pip+numpy-pip+pytest-pip+sympy
· Running 12 total benchmarks (2 commits * 1 environments * 6 benchmarks)
[ 0.00%] · For kda commit 9493049b <master> (round 1/2):
[ 0.00%] ·· Building for conda-py3.11-pip+networkx-pip+numpy-pip+pytest-pip+sympy
[ 0.00%] ·· Benchmarking conda-py3.11-pip+networkx-pip+numpy-pip+pytest-pip+sympy
[ 8.33%] ··· Running (bench_diagrams.DirectionalDiagrams.time_generate_directional_diagrams--).
[16.67%] ··· Running (bench_diagrams.FluxDiagrams.time_generate_flux_diagrams--)..
[25.00%] · For kda commit 5a5a77b0 <improve_dir_edge_generation> (round 1/2):
[25.00%] ·· Building for conda-py3.11-pip+networkx-pip+numpy-pip+pytest-pip+sympy
[25.00%] ·· Benchmarking conda-py3.11-pip+networkx-pip+numpy-pip+pytest-pip+sympy
[33.33%] ··· Running (bench_diagrams.DirectionalDiagrams.time_generate_directional_diagrams--).
[41.67%] ··· Running (bench_diagrams.FluxDiagrams.time_generate_flux_diagrams--)..
[50.00%] · For kda commit 5a5a77b0 <improve_dir_edge_generation> (round 2/2):
[50.00%] ·· Benchmarking conda-py3.11-pip+networkx-pip+numpy-pip+pytest-pip+sympy
[54.17%] ··· ...agrams.peakmem_generate_directional_diagrams                ok
[54.17%] ··· ============== ======= =======
             --               return_edges
             -------------- ---------------
                 graph        True   False
             ============== ======= =======
                3-state      80.4M   80.5M
              Hill-5-state   80.7M   80.9M
              Hill-8-state   80.9M   83.5M
              EmrE-8-state   82.5M    102M
              Max-7-state     152M    818M
             ============== ======= =======

[58.33%] ··· ...lDiagrams.time_generate_directional_diagrams                ok
[58.33%] ··· ============== ============ ============
             --                    return_edges
             -------------- -------------------------
                 graph          True        False
             ============== ============ ============
                3-state      623±100μs    791±100μs
              Hill-5-state   2.82±0.2ms   4.14±0.4ms
              Hill-8-state   23.0±0.4ms    46.7±2ms
              EmrE-8-state    179±8ms      328±20ms
              Max-7-state    8.28±0.04s   12.7±0.07s
             ============== ============ ============

[62.50%] ··· ....FluxDiagrams.peakmem_generate_flux_diagrams                ok
[62.50%] ··· ============== =======
                 graph
             -------------- -------
                3-state      80.3M
              Hill-5-state   80.7M
              Hill-8-state   81.1M
              EmrE-8-state   84.6M
             ============== =======

[66.67%] ··· ...ams.FluxDiagrams.time_generate_flux_diagrams                ok
[66.67%] ··· ============== =============
                 graph
             -------------- -------------
                3-state        809±10μs
              Hill-5-state   2.14±0.06ms
              Hill-8-state    14.9±0.3ms
              EmrE-8-state     152±1ms
             ============== =============

[70.83%] ··· ...alDiagrams.peakmem_generate_partial_diagrams                ok
[70.83%] ··· ============== ======= =======
             --               return_edges
             -------------- ---------------
                 graph        True   False
             ============== ======= =======
                3-state      80.3M   80.3M
              Hill-5-state   80.4M   80.4M
              Hill-8-state   80.6M   80.5M
              EmrE-8-state   80.6M   82.1M
              Max-7-state    81.3M    138M
             ============== ======= =======

[75.00%] ··· ...rtialDiagrams.time_generate_partial_diagrams                ok
[75.00%] ··· ============== ============= ============
             --                    return_edges
             -------------- --------------------------
                 graph           True        False
             ============== ============= ============
                3-state        225±9μs      215±3μs
              Hill-5-state     459±10μs     435±20μs
              Hill-8-state   2.69±0.02ms   2.94±0.2ms
              EmrE-8-state     16.7±1ms    15.3±0.1ms
              Max-7-state     1.07±0.05s    960±10ms
             ============== ============= ============

[75.00%] · For kda commit 9493049b <master> (round 2/2):
[75.00%] ·· Building for conda-py3.11-pip+networkx-pip+numpy-pip+pytest-pip+sympy
[75.00%] ·· Benchmarking conda-py3.11-pip+networkx-pip+numpy-pip+pytest-pip+sympy
[79.17%] ··· ...agrams.peakmem_generate_directional_diagrams                ok
[79.17%] ··· ============== ======= =======
             --               return_edges
             -------------- ---------------
                 graph        True   False
             ============== ======= =======
                3-state      80.4M   80.8M
              Hill-5-state   80.6M   80.8M
              Hill-8-state   80.7M   84.9M
              EmrE-8-state   82.2M    109M
              Max-7-state     150M   1.05G
             ============== ======= =======

[83.33%] ··· ...lDiagrams.time_generate_directional_diagrams                ok
[83.33%] ··· ============== ============ ============
             --                    return_edges
             -------------- -------------------------
                 graph          True        False
             ============== ============ ============
                3-state      1.31±0.1ms   2.49±0.3ms
              Hill-5-state    7.64±2ms     11.4±2ms
              Hill-8-state   80.3±10ms     104±5ms
              EmrE-8-state    478±70ms    785±200ms
              Max-7-state    19.9±0.9s    25.4±0.1s
             ============== ============ ============

[87.50%] ··· ....FluxDiagrams.peakmem_generate_flux_diagrams                ok
[87.50%] ··· ============== =======
                 graph
             -------------- -------
                3-state      80.5M
              Hill-5-state   80.8M
              Hill-8-state   81.1M
              EmrE-8-state   84.7M
             ============== =======

[91.67%] ··· ...ams.FluxDiagrams.time_generate_flux_diagrams                ok
[91.67%] ··· ============== =============
                 graph
             -------------- -------------
                3-state        803±10μs
              Hill-5-state   2.14±0.05ms
              Hill-8-state     15.5±2ms
              EmrE-8-state     151±2ms
             ============== =============

[95.83%] ··· ...alDiagrams.peakmem_generate_partial_diagrams                ok
[95.83%] ··· ============== ======= =======
             --               return_edges
             -------------- ---------------
                 graph        True   False
             ============== ======= =======
                3-state      80.7M   80.9M
              Hill-5-state   80.6M   80.5M
              Hill-8-state   80.4M   80.7M
              EmrE-8-state   80.6M   81.9M
              Max-7-state    81.3M    138M
             ============== ======= =======

[100.00%] ··· ...rtialDiagrams.time_generate_partial_diagrams                ok
[100.00%] ··· ============== ============= =============
              --                     return_edges
              -------------- ---------------------------
                  graph           True         False
              ============== ============= =============
                 3-state        235±10μs      216±2μs
               Hill-5-state     455±30μs      434±3μs
               Hill-8-state   2.72±0.05ms   2.52±0.04ms
               EmrE-8-state    17.4±0.9ms     15.5±2ms
               Max-7-state     1.05±0.03s     961±10ms
              ============== ============= =============

| Change   | Before [9493049b] <master>   | After [5a5a77b0] <improve_dir_edge_generation>   |   Ratio | Benchmark (Parameter)                                                                          |
|----------|------------------------------|--------------------------------------------------|---------|------------------------------------------------------------------------------------------------|
| -        | 1.05G                        | 818M                                             |    0.78 | bench_diagrams.DirectionalDiagrams.peakmem_generate_directional_diagrams('Max-7-state', False) |
| -        | 1.31±0.1ms                   | 623±100μs                                        |    0.48 | bench_diagrams.DirectionalDiagrams.time_generate_directional_diagrams('3-state', True)         |
| -        | 104±5ms                      | 46.7±2ms                                         |    0.45 | bench_diagrams.DirectionalDiagrams.time_generate_directional_diagrams('Hill-8-state', False)   |
| -        | 785±200ms                    | 328±20ms                                         |    0.42 | bench_diagrams.DirectionalDiagrams.time_generate_directional_diagrams('EmrE-8-state', False)   |
| -        | 478±70ms                     | 179±8ms                                          |    0.37 | bench_diagrams.DirectionalDiagrams.time_generate_directional_diagrams('EmrE-8-state', True)    |
| -        | 7.64±2ms                     | 2.82±0.2ms                                       |    0.37 | bench_diagrams.DirectionalDiagrams.time_generate_directional_diagrams('Hill-5-state', True)    |
| -        | 11.4±2ms                     | 4.14±0.4ms                                       |    0.36 | bench_diagrams.DirectionalDiagrams.time_generate_directional_diagrams('Hill-5-state', False)   |
| -        | 2.49±0.3ms                   | 791±100μs                                        |    0.32 | bench_diagrams.DirectionalDiagrams.time_generate_directional_diagrams('3-state', False)        |
| -        | 80.3±10ms                    | 23.0±0.4ms                                       |    0.29 | bench_diagrams.DirectionalDiagrams.time_generate_directional_diagrams('Hill-8-state', True)    |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
``` 
</details>